Provided some new functions to save/load system. Save/load system code has been refactored. Added separate PlayerDataHandler script for save/load system instead doing data collection in PlayerController.cs. Added PlayerDataCollector to player prefab. It is collecting player data for save/load system. Added DataPresistenceManager object to Level Michael.
