+++ Game/Assets/Visual Objects/Levels/Player/Player 1.prefab
@@ -16,6 +16,7 @@ GameObject:
   - component: {fileID: 4637083897867490072}
   - component: {fileID: 5181935550237821176}
   - component: {fileID: 6154229274829694059}
+  - component: {fileID: 7517811549029822662}
   m_Layer: 0
   m_Name: PlayerBody
   m_TagString: Player
@@ -346,52 +347,18 @@ LineRenderer:
     generateLightingData: 0
   m_UseWorldSpace: 1
   m_Loop: 0
---- !u!1 &1065707200974324826
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 7035130004071929136}
-  - component: {fileID: 198101523857815735}
-  m_Layer: 0
-  m_Name: DataPresistenceManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &7035130004071929136
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1065707200974324826}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 2923387096383953752}
-  m_RootOrder: 1
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &198101523857815735
+--- !u!114 &7517811549029822662
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1065707200974324826}
+  m_GameObject: {fileID: 371149642541678525}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 07e1ca369140e4fa6a58983f49a8c9cc, type: 3}
+  m_Script: {fileID: 11500000, guid: 346dc9e8adba64f5997d5676fc989c3d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  fileName: asd
-  timeToWait: 10
 --- !u!1 &6936315482669478508
 GameObject:
   m_ObjectHideFlags: 0
@@ -421,7 +388,6 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 78077141472140128}
-  - {fileID: 7035130004071929136}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
