+++ Game/UserSettings/EditorUserSettings.asset
@@ -18,10 +18,10 @@ EditorUserSettings:
       value: 53540303515759085f5d0d7540760a44104e4d73297b746479784562b5e3313d
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5150575655575d5d0c0a552447215d4443154979747a25667d7a4b32b3b6633a
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 5150575655575d5d0c0a552447215d4443154979747a25667d7a4b32b3b6633a
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 0757060551570f025b57552743700944154e1973747b75637b2d1b60bab66161
